Improvements for MongoDB replica set configuration
* Remove upstream pull request which state is uncertain
  https://github.com/puppetlabs/puppetlabs-mongodb/pull/194
* Move mongorc file creation in openstack class
* Configure admin user in openstack class
* Forbid to configure databases and users from slave nodes
  ('if' condition was added defore executing such commands)
* Remove outdated noop tests
* Remove useless connection validator because it runs
  during replica creation
  Partial-bug: #1475948

Fix method which check existing master mongo:
  Closes-bug: #1482154
It contains a few changes in mongodb module,
but they will be proposed to upstream according to
https://bugs.launchpad.net/fuel/+bug/1475948

+++ deployment/puppet/mongodb/lib/puppet/provider/mongodb_database/mongodb.rb
@@ -7,11 +7,16 @@ Puppet::Type.type(:mongodb_database).provide(:mongodb, :parent => Puppet::Provid
 
   def self.instances
     require 'json'
-    dbs = JSON.parse mongo_eval('printjson(db.getMongo().getDBs())')
+    if mongo_eval('db.isMaster().ismaster').to_s.strip == 'true'
+      dbs = JSON.parse mongo_eval('printjson(db.getMongo().getDBs())')
 
-    dbs['databases'].collect do |db|
-      new(:name   => db['name'],
-          :ensure => :present)
+      dbs['databases'].collect do |db|
+        new(:name   => db['name'],
+            :ensure => :present)
+      end
+    else
+      Puppet.warning 'Impossible to get databases from slave'
+      dbs = []
     end
   end
 
@@ -26,11 +31,15 @@ Puppet::Type.type(:mongodb_database).provide(:mongodb, :parent => Puppet::Provid
   end
 
   def create
-    mongo_eval('db.dummyData.insert({"created_by_puppet": 1})', @resource[:name])
+    if mongo_eval('db.isMaster().ismaster').to_s.strip == 'true'
+      mongo_eval('db.dummyData.insert({"created_by_puppet": 1})', @resource[:name])
+    end
   end
 
   def destroy
-    mongo_eval('db.dropDatabase()', @resource[:name])
+    if mongo_eval('db.isMaster().ismaster').to_s.strip == 'true'
+      mongo_eval('db.dropDatabase()', @resource[:name])
+    end
   end
 
   def exists?
